// packages/contracts/generated/ActivitySession.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ActivityInterval } from "./ActivityInterval";
import type { Id } from "./Id";

export type ActivitySession = { session_id: Id, version: number, intervals: Array<ActivityInterval>, };